Added velocity engine to improve 'tree' generation.
The existing 'tree' classes - *DependencyNodeVisitor - have a horrible
mixture of logic and complex text generation. This is hard to read,
test, maintain, or extend. It is much better to do this work in two
stages. The first is logic - extracting the information that the
destination format requires - and the second is the actual text
generation.

The latter can be handled by a template engine (e.g., velocity)
instead of hand-written java code.

Templates have the additional benefit of being easily reviewed
by people familiar with the format but not java. They can usually
be written to be close to the final format, albeit with the
occasional #foreach etc mixed in. This makes it much easier to
update or extend the generated output.

This patch includes both a concrete implementation (wrapping the
Velocity engine) and an abstraction. That interface isn't perfect
(see the exceptions) but its presence should make it easy for
third-party users to evaluate alternate template engines for their
own needs.

This patch also includes a new 'Style' class since some formats
(e.g. dot) are used to generate images and embedded documents.
In these cases the ability to customize fonts, colors, etc is
important. These values could be provided as individual properties
but using an explicit (and easily updated) class as a carrier
should simply the templates and encourage reuse.