Converted DOTDependencyNodeVisitor to use velocity
NOTE: this builds but fails a few existing tests

This patch implements DOTDependencyNodeVisitor to use a velocity
template instead. It also includes some preliminary work on supporting
other formats - mostly in the new "Indent" velocity directive.

This patch also includes updated examples.

This is stil a work-in-progress since I now want to convert
additional DependencyNodeVisitor classes to use a velocity
template instead. This is partically to improve their maintainability,
partly a way to ensure that the velocity-related classes are
sufficiently complete for most new implementations.