Fully converted to templates
NOTE: not ready for merge - this has introduced a few test failures
and I need to investigate them. I suspect many are due to nothing
but format changes, e.g., using an AtomicInteger to generate an id
instead of using 'hashCode()'

NOTE: I need to double-check the TGF format - the legacy implementation
has a '#' between the nodes and edges and I need to determine if that's
required or just a convenient separator.

Fully converted to velocity templates. Removed 'templates' abstraction -
what are the odds of an additional template library since we can
easily add custom velocity directives?

Updated the 'examples' pages.