Added some documentation, refactored to match it
Added some documentation on velocity templates and overall approach when writing
the format-specific macro libraries.

This prompted a modest refactoring of the existing templates to match the
new documentation. It's nothing too deep, just realizing that the recursive
bits look alike once you remember the rest.

There's a small amount of cleanup by using '$!{key}' instead of explicit #if
and #end statements, and using #else when appropriate. This does have the
downside of leaving the line in place (with an empty value) instead of removing
the line entirely.

NOTE: something that came up when updating the README files - we are
calling this a 'dependency tree' but in fact it's a 'dependency graph' -
this is very clear when using the DOT format to create an image.

I don't know if the original JsonDependencyNodeVisitor filtered out all but
the first reference or if it created duplicate records but I know the
velocity template version will do the latter. The existing behavior should
be maintained.